Extracted Range into its own compilation unit.
