+++ src/appointment_ranges/appointment_ranges_lib.cpp
@@ -4,13 +4,6 @@
 
 using namespace std;
 
-bool operator==(const Range& lhs, const Range& rhs)
-{
-    return lhs.yStart == rhs.yStart &&
-        lhs.yEnd == rhs.yEnd &&
-        lhs.pattern == rhs.pattern;
-}
-
 Pattern getPatternFromState(FreeBusy state)
 {
     switch(state)
@@ -23,27 +16,6 @@ Pattern getPatternFromState(FreeBusy state)
         return Pattern::Empty;
     }
 }
-
-void Ranges::addRange(Range&& newRange)
-{
-    // Check if the previous added pattern is the same.  If so, consolidate.
-    if (!ranges.empty() &&
-        ranges.back().yEnd >= newRange.yStart &&
-        ranges.back().pattern == newRange.pattern)
-    {
-        ranges.back().yEnd = newRange.yEnd;
-    }
-    else
-    {
-        ranges.push_back(newRange);
-    }
-}
-
-const vector<Range>& Ranges::getRanges() const
-{
-    return ranges;
-}
-
 
 #include <iostream> // for cerr tracing
