+++ src/appointment_ranges/range.cpp
@@ -1,0 +1,28 @@
+#include "range.h"
+
+bool operator==(const Range& lhs, const Range& rhs)
+{
+    return lhs.yStart == rhs.yStart &&
+        lhs.yEnd == rhs.yEnd &&
+        lhs.pattern == rhs.pattern;
+}
+
+void Ranges::addRange(Range&& newRange)
+{
+    // Check if the previous added pattern is the same.  If so, consolidate.
+    if (!ranges.empty() &&
+        ranges.back().yEnd >= newRange.yStart &&
+        ranges.back().pattern == newRange.pattern)
+    {
+        ranges.back().yEnd = newRange.yEnd;
+    }
+    else
+    {
+        ranges.push_back(newRange);
+    }
+}
+
+const std::vector<Range>& Ranges::getRanges() const
+{
+    return ranges;
+}
